Fixed some warnings and changed indentations to 2 spaces

+++ orianne/include/orianne/ftp_server.h
@@ -4,15 +4,15 @@
 #include <string>
 
 namespace orianne {
-	
-	class FtpServer {
-	public:
-		FtpServer(boost::asio::io_service& io_service, int port, std::string path);
-		void start();
-		
-	private:
-		std::string path;
-		boost::asio::ip::tcp::acceptor acceptor;
-	};
-	
+
+  class FtpServer {
+  public:
+    FtpServer(boost::asio::io_service& io_service, uint16_t port, std::string path);
+    void start();
+
+  private:
+    std::string path;
+    boost::asio::ip::tcp::acceptor acceptor;
+  };
+
 }
